rename module from happyoct to lisbon.
