+++ app/app.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/evmos/evmos/v10/x/happyoct"
-	happyoctkeeper "github.com/evmos/evmos/v10/x/happyoct/keeper"
-	happyocttypes "github.com/evmos/evmos/v10/x/happyoct/types"
+	"github.com/evmos/evmos/v10/x/lisbon"
+	lisbonkeeper "github.com/evmos/evmos/v10/x/lisbon/keeper"
+	lisbontypes "github.com/evmos/evmos/v10/x/lisbon/types"
 	"io"
 	"net/http"
 	"os"
@@ -205,7 +205,7 @@ var (
 		vesting.AppModuleBasic{},
 		evm.AppModuleBasic{},
 		feemarket.AppModuleBasic{},
-		happyoct.AppModuleBasic{},
+		lisbon.AppModuleBasic{},
 		inflation.AppModuleBasic{},
 		erc20.AppModuleBasic{},
 		incentives.AppModuleBasic{},
@@ -223,7 +223,7 @@ var (
 		govtypes.ModuleName:            {authtypes.Burner},
 		ibctransfertypes.ModuleName:    {authtypes.Minter, authtypes.Burner},
 		evmtypes.ModuleName:            {authtypes.Minter, authtypes.Burner}, // used for secure addition and subtraction of balance using module account
-		happyocttypes.ModuleName:       {authtypes.Minter, authtypes.Burner},
+		lisbontypes.ModuleName:         {authtypes.Minter, authtypes.Burner},
 		inflationtypes.ModuleName:      {authtypes.Minter},
 		erc20types.ModuleName:          {authtypes.Minter, authtypes.Burner},
 		claimstypes.ModuleName:         nil,
@@ -284,7 +284,7 @@ type Evmos struct {
 	FeeMarketKeeper feemarketkeeper.Keeper
 
 	// Evmos keepers
-	HappyOctKeeper   happyoctkeeper.Keeper
+	HappyOctKeeper   lisbonkeeper.Keeper
 	InflationKeeper  inflationkeeper.Keeper
 	ClaimsKeeper     *claimskeeper.Keeper
 	Erc20Keeper      erc20keeper.Keeper
@@ -346,7 +346,7 @@ func NewEvmos(
 		// ethermint keys
 		evmtypes.StoreKey, feemarkettypes.StoreKey,
 		// evmos keys
-		happyocttypes.StoreKey,
+		lisbontypes.StoreKey,
 		inflationtypes.StoreKey, erc20types.StoreKey, incentivestypes.StoreKey,
 		epochstypes.StoreKey, claimstypes.StoreKey, vestingtypes.StoreKey,
 		revenuetypes.StoreKey,
@@ -447,7 +447,7 @@ func NewEvmos(
 	)
 
 	// Evmos Keeper
-	app.HappyOctKeeper = happyoctkeeper.NewKeeper(keys[happyocttypes.StoreKey], appCodec, app.GetSubspace(happyocttypes.ModuleName),
+	app.HappyOctKeeper = lisbonkeeper.NewKeeper(keys[lisbontypes.StoreKey], appCodec, app.GetSubspace(lisbontypes.ModuleName),
 		app.BankKeeper, &stakingKeeper, app.AccountKeeper, app.EvmKeeper, authtypes.FeeCollectorName)
 
 	app.InflationKeeper = inflationkeeper.NewKeeper(
@@ -613,7 +613,7 @@ func NewEvmos(
 		evm.NewAppModule(app.EvmKeeper, app.AccountKeeper),
 		feemarket.NewAppModule(app.FeeMarketKeeper),
 		// Evmos app modules
-		happyoct.NewAppModule(app.HappyOctKeeper, app.AccountKeeper),
+		lisbon.NewAppModule(app.HappyOctKeeper, app.AccountKeeper),
 		inflation.NewAppModule(app.InflationKeeper, app.AccountKeeper, app.StakingKeeper),
 		erc20.NewAppModule(app.Erc20Keeper, app.AccountKeeper),
 		incentives.NewAppModule(app.IncentivesKeeper, app.AccountKeeper),
@@ -638,7 +638,7 @@ func NewEvmos(
 		feemarkettypes.ModuleName,
 		evmtypes.ModuleName,
 		revenuetypes.ModuleName,
-		happyocttypes.ModuleName,
+		lisbontypes.ModuleName,
 		slashingtypes.ModuleName,
 		evidencetypes.ModuleName,
 		stakingtypes.ModuleName,
@@ -677,7 +677,7 @@ func NewEvmos(
 		capabilitytypes.ModuleName,
 		authtypes.ModuleName,
 		banktypes.ModuleName,
-		happyocttypes.ModuleName,
+		lisbontypes.ModuleName,
 		slashingtypes.ModuleName,
 		genutiltypes.ModuleName,
 		evidencetypes.ModuleName,
@@ -725,7 +725,7 @@ func NewEvmos(
 		upgradetypes.ModuleName,
 		// Evmos modules
 		vestingtypes.ModuleName,
-		happyocttypes.ModuleName,
+		lisbontypes.ModuleName,
 		inflationtypes.ModuleName,
 		erc20types.ModuleName,
 		incentivestypes.ModuleName,
